fst: Handle EditIdle
When running vst2 JUCE gui in Carla on Linux - the window is always
black and never updated. When GUI is opened JUCE expects that hosts
would drive event loop with effEditIdle messages.

The only event that plugin continuously receives has opcode 19,
assigning effEditIdle this opcode results in a correct rendering of
the plugin's window.

Therefor I strongly believe that effEditIdle = 19